requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0
argparse==1.2.1
django-bootstrap3==4.7.0
django-bootstrap3-datetimepicker==2.2.3
django-ckeditor-updated==4.4.0
django-parsley==0.5
django-simple-captcha==0.4.4
django-widget-tweaks==1.3
enum==0.4.4
lxml==3.4.2
pip==1.5.6
setuptools==3.6
six==1.9.0
wsgiref==0.1.2